fix: Move Location import from api to model to match 0.5.3
tsia
