+++ homeassistant/components/canary/sensor.py
@@ -3,8 +3,7 @@ from __future__ import annotations
 
 from typing import Final
 
-from canary.api import Location
-from canary.model import Device, SensorType
+from canary.model import Device, Location, SensorType
 
 from homeassistant.components.sensor import SensorDeviceClass, SensorEntity
 from homeassistant.config_entries import ConfigEntry
